function/requirements.txt
functions-framework==3.*
cloudevents==1.11.0
google-cloud-storage>=2.16.0
google-cloud-aiplatform>=1.67.0